Moved the Button component into the reusable component folder
